+++ lib/controller/group_chat_controller.dart
@@ -1,0 +1,6 @@
+import 'package:get/get.dart';
+
+class GroupChatController extends GetxController {
+  RxInt selectedCategoryIndex = 999.obs;
+  RxInt selectedParticipantsNumberIndex = 999.obs;
+}
